refactor(#186): SIP-0097 slice 5 — transport moves to TaskDispatcher, interim callables retired
Extract the request/reply transport verbatim into
adapters/cycles/task_dispatcher.py (SIP-0097 §6.1): dispatch_task /
dispatch_with_retry / create_task_run_if_enabled plus the internal
_publish_and_await / _task_heartbeat / _start_task_activity /
_finish_task_activity. Per-task observability (RuntimeActivity, Prefect
task-run lifecycle, correlation-context scoping) starts and finishes in
the dispatcher per the observability ownership rule.

AC#9: CorrectionRunner and PulseBoundaryRunner now take the
TaskDispatcher itself — the slice-3/4 interim executor-supplied dispatch
callables no longer exist; the dependency graph matches §6. The
store_artifact callable stays executor-supplied (§6.7 residual).

Routing stays with the executor (§6.1): dispatch_with_retry receives
_handle_task_outcome as a per-call closure bound over the sequential
loop's orchestration state (loop-scoped ints bind as defaults — fixed at
definition time exactly like the old parameter passing) and only acts on
the returned action token. The dispatcher takes the event bus for the
retry loop's per-attempt TASK_* emits (§6.1's dependency list omits it,
but its named method needs it — same events, same ordering). The §6.1
cancellation probe and LLM observability deps are deliberately NOT wired:
the transport never checked cancellation (orchestration loops do, between
tasks) and per-task generation traces are agent-side today —
defer-infra-completeness; both disclosed as deviations from §6.1's
dependency list.

Test split (§9) + mechanical updates (disclosed per §7.7):
- TestDispatchTask / TestDispatchTaskPrefectLifecycle /
  TestPublishAndAwaitInvariants move to test_task_dispatcher.py and
  construct the dispatcher directly (no executor) — assertions unchanged;
  patch targets re-pointed to the new module.
- test_executor_activity_instrumentation.py → test_task_dispatcher_activity.py
  (git mv): same assertions against a directly-constructed TaskDispatcher.
- Correction/pulse standalone tests swap the two dispatch callables for a
  _PassthroughDispatcher stand-in; propagation/pulse tests stub
  ex._task_dispatcher.dispatch_task instead of ex._dispatch_task.
- Emission guardrails re-anchored: TaskDispatcher owns the retry loop's 2
  TASK_* emit sites; executor 19→17 emits (refs unchanged at 12 — fan-out
  still references all three TASK_* types); total unchanged at 41.

All moved methods pass C901=12 at their new home — no new per-file
ignore; pyproject.toml untouched. Executor: 2,036 → 1,727 lines (3,358
at arc start).
