+++ adapters/cycles/dispatched_flow_executor.py
@@ -14,7 +14,6 @@ from __future__ import annotations
 
 import asyncio
 import dataclasses
-import json
 import logging
 import os
 import time
@@ -32,6 +31,7 @@ from adapters.cycles.execution_errors import (
 )
 from adapters.cycles.pulse_boundary_runner import PulseBoundaryRunner
 from adapters.cycles.run_completion import RunCompletion, resolve_terminal_outcome
+from adapters.cycles.task_dispatcher import TaskDispatcher
 from adapters.cycles.task_naming import build_task_name
 from squadops.cycles.agent_config import build_agent_resolver
 from squadops.cycles.checkpoint import RunCheckpoint
@@ -42,11 +42,10 @@ from squadops.cycles.task_outcome import TaskOutcome
 from squadops.cycles.task_plan import generate_task_plan
 from squadops.events.types import EventType
 from squadops.ports.cycles.flow_execution import FlowExecutionPort
-from squadops.runtime import reasons
 from squadops.runtime.admission import admit_participants, release_participants
 from squadops.runtime.recruitment import reserve_buffer_decision
 from squadops.tasks.models import TaskEnvelope, TaskResult
-from squadops.telemetry.context import use_correlation_context, use_run_ids
+from squadops.telemetry.context import use_correlation_context
 from squadops.telemetry.models import CorrelationContext
 
 if TYPE_CHECKING:
@@ -95,6 +94,7 @@ class DispatchedFlowExecutor(FlowExecutionPort):
         run_completion: RunCompletion | None = None,
         correction_runner: CorrectionRunner | None = None,
         pulse_boundary_runner: PulseBoundaryRunner | None = None,
+        task_dispatcher: TaskDispatcher | None = None,
     ) -> None:
         self._cycle_registry = cycle_registry
         self._artifact_vault = artifact_vault
@@ -138,41 +138,36 @@ class DispatchedFlowExecutor(FlowExecutionPort):
 
             event_bus = NoOpCycleEventBus()
         self._cycle_event_bus = event_bus
-        # SIP-0097 §6.3: correction-protocol collaborator. Plain injected
-        # collaborator with a default composed from this executor's own deps.
-        # The three callables are the sanctioned interim seam (slice 3 <
-        # slice 5): dispatch_task/create_task_run are re-pointed to
-        # TaskDispatcher in slice 5 (AC#9); store_artifact stays
-        # executor-supplied (artifact plumbing is §6.7 residual). Late-bound
-        # lambdas (not captured bound methods) so the runner always sees the
-        # executor's *current* dispatch — tests stub `_dispatch_task` on the
-        # instance.
+        # SIP-0097 §6.1: task-dispatch collaborator (request/reply transport +
+        # per-task observability). Composed first — the correction and pulse
+        # runners below depend on it (the §6 final dependency graph; the
+        # slice-3/4 interim dispatch callables are retired per AC#9).
+        self._task_dispatcher = task_dispatcher or TaskDispatcher(
+            queue=queue,
+            reply_router=reply_router,
+            workflow_tracker=workflow_tracker,
+            activity_port=activity_port,
+            event_bus=event_bus,
+            task_timeout=task_timeout,
+        )
+        # SIP-0097 §6.3: correction-protocol collaborator. store_artifact stays
+        # an executor-supplied late-bound callable (artifact plumbing is §6.7
+        # executor residual, residual-but-watched).
         self._correction_runner = correction_runner or CorrectionRunner(
             cycle_registry=cycle_registry,
             artifact_vault=artifact_vault,
             event_bus=event_bus,
-            dispatch_task=lambda envelope, run_id, **kw: self._dispatch_task(
-                envelope, run_id, **kw
-            ),
-            create_task_run=lambda flow_run_id, envelope: self._create_task_run_if_enabled(
-                flow_run_id, envelope
-            ),
+            task_dispatcher=self._task_dispatcher,
             store_artifact=lambda *args, **kw: self._store_artifact(*args, **kw),
         )
         # SIP-0097 §6.2: pulse-boundary collaborator (boundary verification +
-        # bounded repair loop). Same interim callable seam as the correction
-        # runner above; LLM observability is event-emission-only per the
-        # observability ownership rule.
+        # bounded repair loop). LLM observability is event-emission-only per
+        # the observability ownership rule.
         self._pulse_boundary_runner = pulse_boundary_runner or PulseBoundaryRunner(
             cycle_registry=cycle_registry,
             event_bus=event_bus,
             llm_observability=llm_observability,
-            dispatch_task=lambda envelope, run_id, **kw: self._dispatch_task(
-                envelope, run_id, **kw
-            ),
-            create_task_run=lambda flow_run_id, envelope: self._create_task_run_if_enabled(
-                flow_run_id, envelope
-            ),
+            task_dispatcher=self._task_dispatcher,
             store_artifact=lambda *args, **kw: self._store_artifact(*args, **kw),
         )
         # SIP-0097 §6.6: per-run pulse summaries live on the RunLedger created
@@ -326,7 +321,7 @@ class DispatchedFlowExecutor(FlowExecutionPort):
             # logs emitted between dispatches (workload progression, gate
             # decisions, executor lifecycle) carry flow_run_id (and no
             # task_run_id) and land in the flow-run pane in the Prefect UI.
-            # _dispatch_task nests its own per-task scope inside this one.
+            # TaskDispatcher.dispatch_task nests its own per-task scope inside this one.
             flow_ctx = CorrelationContext(cycle_id=cycle_id, flow_run_id=flow_run_id)
             with use_correlation_context(flow_ctx):
                 logger.info(
@@ -706,248 +701,6 @@ class DispatchedFlowExecutor(FlowExecutionPort):
         return await self._cycle_registry.create_run(next_run)
 
     # ------------------------------------------------------------------
-    # Dispatch
-    # ------------------------------------------------------------------
-
-    # SIP-0087: task-run lifecycle lives here (moved out of WorkflowTrackerBridge) so
-    # the task_run_id is known before the agent starts producing logs.
-    async def _create_task_run_if_enabled(
-        self,
-        flow_run_id: str | None,
-        envelope: TaskEnvelope,
-    ) -> str | None:
-        """Create a Prefect task_run + set RUNNING. Returns task_run_id or None."""
-        if self._workflow_tracker is None or not flow_run_id:
-            return None
-        task_name = build_task_name(envelope)
-        task_run_id = await self._workflow_tracker.create_task_run(
-            flow_run_id, envelope.task_id, task_name
-        )
-        await self._workflow_tracker.set_task_run_state(task_run_id, "RUNNING", "Running")
-        return task_run_id
-
-    async def _task_heartbeat(
-        self,
-        envelope: TaskEnvelope,
-        *,
-        interval: float,
-    ) -> None:
-        """Log a heartbeat line every ``interval`` seconds until cancelled.
-
-        Runs inside the task's ``use_run_ids`` scope (contextvars copy across
-        ``asyncio.create_task``), so emitted records carry the active
-        flow_run_id + task_run_id and land in the right Prefect pane.
-        """
-        start = time.monotonic()
-        capability_id = (
-            envelope.metadata.get("capability_id", envelope.task_type)
-            if envelope.metadata
-            else envelope.task_type
-        )
-        while True:
-            await asyncio.sleep(interval)
-            elapsed = time.monotonic() - start
-            logger.info(
-                "task_heartbeat elapsed=%.1fs capability_id=%s task_id=%s",
-                elapsed,
-                capability_id,
-                envelope.task_id,
-            )
-
-    async def _dispatch_task(
-        self,
-        envelope: TaskEnvelope,
-        run_id: str,
-        *,
-        flow_run_id: str | None = None,
-        task_run_id: str | None = None,
-        heartbeat_interval: float = 30.0,
-    ) -> TaskResult:
-        """Publish task to agent queue, wait for result on reply queue.
-
-        If ``task_run_id`` is not supplied but ``flow_run_id`` is and a
-        :class:`WorkflowTrackerPort` is wired, one is created here — supports
-        correction/repair paths that don't pre-create the workflow run.
-
-        Enters a ``CorrelationContext`` scope with flow/task run IDs so the
-        ``PrefectLogHandler`` scopes handler logs to the right Prefect task
-        pane, and spawns a periodic heartbeat coroutine so long-running LLM
-        calls show liveness in the UI.
-        """
-        if task_run_id is None:
-            task_run_id = await self._create_task_run_if_enabled(flow_run_id, envelope)
-
-        # SIP-0087 B1: propagate run IDs to the agent over the wire so the
-        # agent's PrefectLogHandler can scope handler logs to the right task
-        # pane. The agent enters use_correlation_context(...) on receipt.
-        envelope = dataclasses.replace(
-            envelope,
-            flow_run_id=flow_run_id or "",
-            task_run_id=task_run_id or "",
-        )
-
-        base_ctx = CorrelationContext.from_envelope(
-            envelope,
-            agent_id=envelope.agent_id or "",
-            agent_role=(envelope.metadata.get("role") if envelope.metadata else None),
-        )
-
-        with (
-            use_correlation_context(base_ctx),
-            use_run_ids(flow_run_id=flow_run_id, task_run_id=task_run_id),
-        ):
-            heartbeat = asyncio.create_task(
-                self._task_heartbeat(envelope, interval=heartbeat_interval),
-                name=f"prefect-heartbeat-{envelope.task_id}",
-            )
-            # SIP-0089 §4.4: open a RuntimeActivity for this task (best-effort).
-            activity_id = await self._start_task_activity(envelope)
-            try:
-                result = await self._publish_and_await(envelope, run_id)
-            except BaseException:
-                # Reply wait raised (rare — _publish_and_await usually returns a
-                # FAILED TaskResult): record the task activity as failed.
-                await self._finish_task_activity(activity_id, None)
-                raise
-            else:
-                await self._finish_task_activity(activity_id, result)
-                return result
-            finally:
-                heartbeat.cancel()
-                try:
-                    await heartbeat
-                except asyncio.CancelledError:
-                    pass
-
-    async def _start_task_activity(self, envelope: TaskEnvelope) -> str | None:
-        """Open a RuntimeActivity for a dispatched cycle task (§4.4, executor-side).
-
-        Best-effort: returns the activity id, or None when disabled/failed — it
-        must never raise, so observability can't break dispatch. D9 (one active
-        activity per agent) is enforced by the partial unique index; a leftover
-        active row from a prior crashed task makes this conflict, which we swallow.
-        """
-        if self._activity_port is None:
-            return None
-        try:
-            activity = await self._activity_port.start_activity(
-                envelope.agent_id,
-                mode="cycle",
-                activity_type=envelope.task_type,
-                goal=envelope.task_name or f"{envelope.task_type} ({envelope.cycle_id})",
-                source_kind="cycle_task",
-                source_ref=envelope.task_id,
-                cycle_id=envelope.cycle_id,
-                task_id=envelope.task_id,
-            )
-            return activity.runtime_activity_id
-        except Exception:
-            logger.warning(
-                "best-effort start_activity failed for task %s", envelope.task_id, exc_info=True
-            )
-            return None
-
-    async def _finish_task_activity(
-        self, activity_id: str | None, result: TaskResult | None
-    ) -> None:
-        """Terminate a task's RuntimeActivity (§4.4, executor-side, best-effort).
-
-        SUCCEEDED → complete; anything else (FAILED/CANCELED/raised/None) → fail.
-        Never raises. `update_state`'s active-only guard makes this safe even if
-        the activity was already terminalized elsewhere.
-        """
-        if self._activity_port is None or activity_id is None:
-            return
-        try:
-            if result is not None and result.status == "SUCCEEDED":
-                await self._activity_port.complete_activity(activity_id)
-            else:
-                reason = (result.error if result is not None else None) or reasons.ACTIVITY_FAILED
-                await self._activity_port.fail_activity(activity_id, reason)
-        except Exception:
-            logger.warning("best-effort finish_activity failed for %s", activity_id, exc_info=True)
-
-    async def _publish_and_await(
-        self,
-        envelope: TaskEnvelope,
-        run_id: str,
-    ) -> TaskResult:
-        """Dispatch over ``{agent_id}_comms`` and await the reply via the router.
-
-        SIP-0094: replaces the per-run ``cycle_results_{run_id}`` polling loop
-        with a long-lived per-agent subscription. The reply for ``task_id`` is
-        delivered through :class:`ReplyRouter`, which resolves the future this
-        method awaits.
-
-        Invariants:
-        - **Ordering (D14/#2):** ``ensure_subscribed → register → publish`` — the
-          consumer is live before any reply can arrive (no first-dispatch race).
-        - **No pending-future leak (#9):** every exit path — success, timeout,
-          publish failure, router/await failure, cancellation — removes
-          ``task_id`` from the router so it never lingers.
-        """
-        if self._reply_router is None:
-            raise RuntimeError(
-                "DispatchedFlowExecutor requires a ReplyRouter to dispatch (SIP-0094)"
-            )
-
-        reply_queue = f"{envelope.agent_id}_replies"
-        queue_name = f"{envelope.agent_id}_comms"
-
-        # Open the agent's reply subscription and register our future BEFORE
-        # publishing, so a fast reply can't arrive before we're listening.
-        await self._reply_router.ensure_subscribed(envelope.agent_id)
-        fut = self._reply_router.register(envelope.task_id)
-
-        message = {
-            "action": "comms.task",
-            "metadata": {
-                "reply_queue": reply_queue,
-                "correlation_id": envelope.correlation_id,
-            },
-            "payload": envelope.to_dict(),
-        }
-
-        # #10: if publish fails after register(), the agent never got the task —
-        # drop the pending future so it doesn't leak.
-        try:
-            await self._queue.publish(queue_name, json.dumps(message))
-        except Exception:
-            self._reply_router.cancel(envelope.task_id)
-            raise
-
-        logger.info(
-            "Dispatched task %s (%s) to %s, awaiting reply on %s",
-            envelope.task_id,
-            envelope.task_type,
-            queue_name,
-            reply_queue,
-        )
-
-        try:
-            return await asyncio.wait_for(fut, timeout=self._task_timeout)
-        except asyncio.CancelledError:
-            self._reply_router.cancel(envelope.task_id)
-            raise
-        except TimeoutError:
-            self._reply_router.cancel(envelope.task_id)
-            return TaskResult(
-                task_id=envelope.task_id,
-                status="FAILED",
-                error=f"Timed out waiting for agent {envelope.agent_id} after {self._task_timeout}s",
-            )
-        except Exception as exc:
-            # Router-side failure surfaced via the future (e.g. a malformed
-            # reply that failed TaskResult.from_dict, or ReplyRouterStopped on
-            # shutdown). The future is already settled; just guard the leak.
-            self._reply_router.cancel(envelope.task_id)
-            return TaskResult(
-                task_id=envelope.task_id,
-                status="FAILED",
-                error=f"Reply wait for agent {envelope.agent_id} failed: {exc}",
-            )
-
-    # ------------------------------------------------------------------
     # Execution strategies
     # ------------------------------------------------------------------
 
@@ -1144,7 +897,9 @@ class DispatchedFlowExecutor(FlowExecutionPort):
             # SIP-0087: executor owns Prefect task-run creation so task_run_id
             # is available for contextvar scoping + heartbeat before the
             # agent starts emitting logs.
-            task_run_id = await self._create_task_run_if_enabled(flow_run_id, envelope)
+            task_run_id = await self._task_dispatcher.create_task_run_if_enabled(
+                flow_run_id, envelope
+            )
 
             # SIP-0077: task.dispatched event (bridge now only reads terminal
             # state from context; creation already happened above).
@@ -1164,23 +919,43 @@ class DispatchedFlowExecutor(FlowExecutionPort):
                 },
             )
 
-            # Dispatch + retry loop for retryable failures (SIP-0079)
-            task_succeeded, result = await self._dispatch_with_retry(
-                enriched=enriched,
-                envelope=envelope,
-                cycle=cycle,
-                run_id=run_id,
-                task_attempt_counts=task_attempt_counts,
-                consecutive_failures=consecutive_failures,
-                correction_attempts=correction_attempts,
-                prior_outputs=prior_outputs,
-                all_artifact_refs=all_artifact_refs,
-                stored_artifacts=stored_artifacts,
-                completed_task_ids=completed_task_ids,
-                plan_delta_refs=plan_delta_refs,
-                profile=profile,
+            # Dispatch + retry loop for retryable failures (SIP-0079).
+            # Routing stays here (§6.1): the dispatcher receives the outcome
+            # decision as a closure over this loop's orchestration state and
+            # only acts on the returned action token.
+            # Loop-scoped values bind as defaults: fixed at definition time,
+            # exactly like the old per-call parameter passing (and B023-safe).
+            async def _route_outcome(
+                result,
+                _envelope=envelope,
+                _consecutive_failures=consecutive_failures,
+                _correction_attempts=correction_attempts,
+            ):
+                return await self._handle_task_outcome(
+                    result=result,
+                    envelope=_envelope,
+                    cycle=cycle,
+                    run_id=run_id,
+                    task_attempt_counts=task_attempt_counts,
+                    consecutive_failures=_consecutive_failures,
+                    correction_attempts=_correction_attempts,
+                    prior_outputs=prior_outputs,
+                    all_artifact_refs=all_artifact_refs,
+                    stored_artifacts=stored_artifacts,
+                    completed_task_ids=completed_task_ids,
+                    plan_delta_refs=plan_delta_refs,
+                    profile=profile,
+                    flow_run_id=flow_run_id,
+                )
+
+            task_succeeded, result = await self._task_dispatcher.dispatch_with_retry(
+                enriched,
+                envelope,
+                cycle,
+                run_id,
                 flow_run_id=flow_run_id,
                 task_run_id=task_run_id,
+                handle_task_outcome=_route_outcome,
             )
 
             if not task_succeeded:
@@ -1350,7 +1125,6 @@ class DispatchedFlowExecutor(FlowExecutionPort):
         stored_artifacts: list[tuple[str, ArtifactRef]],
     ) -> TaskEnvelope:
         """Build chain context inputs and return an enriched envelope."""
-        import dataclasses
 
         extra_inputs: dict[str, Any] = {
             "prior_outputs": prior_outputs,
@@ -1373,90 +1147,6 @@ class DispatchedFlowExecutor(FlowExecutionPort):
                 **extra_inputs,
             },
         )
-
-    async def _dispatch_with_retry(
-        self,
-        enriched: TaskEnvelope,
-        envelope: TaskEnvelope,
-        cycle: Cycle,
-        run_id: str,
-        task_attempt_counts: dict[str, int],
-        consecutive_failures: int,
-        correction_attempts: int,
-        prior_outputs: dict[str, Any],
-        all_artifact_refs: list[str],
-        stored_artifacts: list[tuple[str, ArtifactRef]],
-        completed_task_ids: list[str],
-        plan_delta_refs: list[str],
-        profile: Any = None,
-        flow_run_id: str | None = None,
-        task_run_id: str | None = None,
-    ) -> tuple[bool, TaskResult]:
-        """Dispatch a task with retry loop for retryable failures.
-
-        Returns (task_succeeded, result). Raises on unrecoverable failures.
-        """
-        while True:
-            result = await self._dispatch_task(
-                enriched,
-                run_id,
-                flow_run_id=flow_run_id,
-                task_run_id=task_run_id,
-            )
-
-            # SIP-0087: task_run_id carried in context so the bridge can set
-            # terminal Prefect state without owning the ID.
-            task_context = {
-                "cycle_id": cycle.cycle_id,
-                "run_id": run_id,
-                "task_run_id": task_run_id or "",
-            }
-            if result.status == "SUCCEEDED":
-                self._cycle_event_bus.emit(
-                    EventType.TASK_SUCCEEDED,
-                    entity_type="task",
-                    entity_id=envelope.task_id,
-                    context=task_context,
-                    payload={"task_type": envelope.task_type},
-                )
-            else:
-                self._cycle_event_bus.emit(
-                    EventType.TASK_FAILED,
-                    entity_type="task",
-                    entity_id=envelope.task_id,
-                    context=task_context,
-                    payload={
-                        "task_type": envelope.task_type,
-                        "error": result.error or "",
-                    },
-                )
-
-            if result.status != "SUCCEEDED":
-                action = await self._handle_task_outcome(
-                    result=result,
-                    envelope=envelope,
-                    cycle=cycle,
-                    run_id=run_id,
-                    task_attempt_counts=task_attempt_counts,
-                    consecutive_failures=consecutive_failures,
-                    correction_attempts=correction_attempts,
-                    prior_outputs=prior_outputs,
-                    all_artifact_refs=all_artifact_refs,
-                    stored_artifacts=stored_artifacts,
-                    completed_task_ids=completed_task_ids,
-                    plan_delta_refs=plan_delta_refs,
-                    profile=profile,
-                    flow_run_id=flow_run_id,
-                )
-
-                if action == "continue":
-                    continue  # retry same task
-                elif action == "break_correction":
-                    return False, result
-
-                # "raise" actions handled inside _handle_task_outcome
-
-            return True, result
 
     async def _restore_checkpoint_state(
         self,
@@ -1775,7 +1465,6 @@ class DispatchedFlowExecutor(FlowExecutionPort):
         flow_run_id: str | None = None,
     ) -> None:
         """Fan-out/fan-in: dispatch all tasks concurrently, await all."""
-        import dataclasses
 
         if await self._is_cancelled(run_id):
             raise _CancellationError(run_id)
@@ -1785,7 +1474,9 @@ class DispatchedFlowExecutor(FlowExecutionPort):
         for envelope in plan:
             # SIP-0087: create Prefect task_run here so task_run_id is in the
             # TASK_DISPATCHED context and visible to the contextvar scope.
-            task_run_id = await self._create_task_run_if_enabled(flow_run_id, envelope)
+            task_run_id = await self._task_dispatcher.create_task_run_if_enabled(
+                flow_run_id, envelope
+            )
             task_run_ids.append(task_run_id)
 
             self._cycle_event_bus.emit(
@@ -1808,7 +1499,7 @@ class DispatchedFlowExecutor(FlowExecutionPort):
                 inputs={**envelope.inputs, "prior_outputs": {}, "artifact_refs": []},
             )
             tasks.append(
-                self._dispatch_task(
+                self._task_dispatcher.dispatch_task(
                     enriched,
                     run_id,
                     flow_run_id=flow_run_id,
